move session inside agent folder
